Add auth components and mobile styling
